up api msme

+++ SME_API_MSME/SME_API_MSME/Entities/TBudgetPlan.cs
@@ -13,7 +13,7 @@ public partial class TBudgetPlan
 
     public string? ItemActivityDetail { get; set; }
 
-    public decimal? ActivityBudget { get; set; }
+    public double? ActivityBudget { get; set; }
 
     public string? ExpenseTypeName { get; set; }
