+++ SME_API_MSME/SME_API_MSME/Entities/TProjectActivity.cs
@@ -13,9 +13,9 @@ public partial class TProjectActivity
 
     public string? ItemActivityDetail { get; set; }
 
-    public decimal? ActivityWeight { get; set; }
+    public double? ActivityWeight { get; set; }
 
-    public int? Target { get; set; }
+    public double? Target { get; set; }
 
     public string? UnitName { get; set; }
